refactor: 장소 상세페이지 컴포넌트들 분리
