+++ src/components/domain/Place/PlaceSlider.tsx
