+++ src/components/domain/Place/RelevantCourses.tsx
@@ -1,43 +1,38 @@
-import CourseItem from '~/components/common/CourseList/CourseItem';
-import { ICourseItem } from '~/types/course';
-import SliderContainer from '~/components/common/SliderContainer';
-import { useState } from 'react';
-import ConfirmModal from '~/components/common/ConfirmModal';
-import { useRouter } from 'next/router';
 import styled from '@emotion/styled';
+import { Link, Text, Title } from '~/components/atom';
+import theme from '~/styles/theme';
+import { ICourseItem } from '~/types/course';
+import RelevantSlider from './RelevantSlider';
 
-interface SliderProps {
+interface RelevantCoursesProps {
+  placeId: number;
   courses: ICourseItem[];
 }
-const RelevantCourses = ({ courses }: SliderProps) => {
-  const router = useRouter();
-  const [modalVisible, setModalVisible] = useState(false);
 
-  const handleGoLogin = () => {
-    setModalVisible(false);
-    router.push('/login');
-  };
-
+const RelevantCourses = ({ placeId, courses }: RelevantCoursesProps) => {
   return (
-    <>
-      <StyledSlider itemMargin={10}>
-        {courses.map((course) => (
-          <CourseItem course={course} key={course.id} grid={1} onModal={setModalVisible} />
-        ))}
-      </StyledSlider>
-      <ConfirmModal
-        visible={modalVisible}
-        onClose={() => setModalVisible(false)}
-        onConfirm={handleGoLogin}
-        message="로그인이 필요한 서비스입니다."
-        subMessage="로그인 페이지로 이동할까요?"
-      />
-    </>
+    <Container>
+      <RelevantHeader>
+        <Title size="sm">이 장소가 포함된 코스</Title>
+        <Link href={`/course/search/${placeId}`}>
+          <Text color="gray">전체보기</Text>
+        </Link>
+      </RelevantHeader>
+      <RelevantSlider courses={courses} />
+    </Container>
   );
 };
 
 export default RelevantCourses;
 
-const StyledSlider = styled(SliderContainer)`
-  height: 360px;
+const Container = styled.div`
+  margin-top: 30px;
+  padding: 50px 0;
+  border-top: 1px solid ${theme.color.borderGray};
+`;
+
+const RelevantHeader = styled.div`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
 `;
