+++ src/components/domain/Place/index.tsx
@@ -1,0 +1,6 @@
+export { default as ToggleMap } from './ToggleMap';
+export { default as PlaceSlider } from './PlaceSlider';
+export { default as PostDescription } from './PostDescription';
+export { default as PostHeader } from './PostHeader';
+export { default as RelevantCourses } from './RelevantCourses';
+export { default as RelevantSlider } from './RelevantSlider';
