+++ src/pages/place/[id].tsx
@@ -1,19 +1,20 @@
-import styled from '@emotion/styled';
-import { NextPageContext } from 'next';
-import Head from 'next/head';
 import React, { useEffect, useState } from 'react';
-import { Button, Icon, Link, PageContainer, Text, Title } from '~/components/atom';
+import styled from '@emotion/styled';
+import theme from '~/styles/theme';
+import { NextPageContext } from 'next';
+import { PageContainer } from '~/components/atom';
+import PageHead from '~/components/common/PageHead';
+import {
+  ToggleMap,
+  PostDescription,
+  PostHeader,
+  RelevantCourses,
+  PlaceSlider
+} from '~/components/domain/Place';
 import Comment from '~/components/common/Comment';
 import DetailSidebar from '~/components/common/DetailSidebar';
-import ImageViewer from '~/components/common/ImageViewer';
-import PageHead from '~/components/common/PageHead';
-import SliderContainer from '~/components/common/SliderContainer';
-import PlaceMap from '~/components/domain/Map/PlaceMap';
-import RelevantCourses from '~/components/domain/Place/RelevantCourses';
-import PlaceSlider from '~/components/domain/PlaceSlider';
 import { useUser } from '~/hooks/useUser';
 import { CourseApi, PlaceApi } from '~/service';
-import theme from '~/styles/theme';
 import { PlacePost } from '~/types';
 import { ICourseItem } from '~/types/course';
 
@@ -57,7 +58,6 @@ const PlaceDetailByPostId = ({ place, placeId, courses }: Props) => {
   const [relevantCourses, setRelevantCourses] = useState(courses || []);
   const { isLoggedIn } = useUser();
   const [detailData, setDetailData] = useState<PlacePost | null>(place);
-  const [isOpenMap, setIsOpenMap] = useState(false);
 
   const getDetailInfo = async () => {
     const places = await PlaceApi.read(placeId);
@@ -88,44 +88,22 @@ const PlaceDetailByPostId = ({ place, placeId, courses }: Props) => {
       <PageHead title={place.name} />
       <main>
         <Container type="detail" style={{ position: 'relative' }}>
-          <PostHeader>
-            <Title level={1} size="lg" fontWeight={700} block>
-              {detailData.name}
-            </Title>
-            <Text size="md" block color="gray">
-              {detailData.roadAddressName || detailData.addressName}
-            </Text>
-          </PostHeader>
-          <Description size="lg" block>
-            <Text color="main" size="lg" fontWeight={700}>
-              {detailData.usedCount}개의 여행코스
-            </Text>
-            에 포함된 장소입니다.
-          </Description>
-          <MapButton width="140px" size="sm" onClick={() => setIsOpenMap(!isOpenMap)}>
-            지도보기{' '}
-            {isOpenMap ? <Icon name="arrowDown" rotate={180} /> : <Icon name="arrowDown" />}
-          </MapButton>
-          {detailData && isOpenMap && (
-            <PlaceMap
-              placeId={Number(detailData.kakaoMapId)}
-              placeType={detailData.category}
-              placeName={detailData.name}
-              center={{ lat: Number(detailData.latitude), lng: Number(detailData.longitude) }}
-            />
-          )}
-
+          <PostHeader
+            name={detailData.name}
+            address={detailData.roadAddressName || detailData.addressName}
+          />
+          <PostDescription usedCount={detailData.usedCount} />
+          <ToggleMap
+            placeId={Number(detailData.kakaoMapId)}
+            placeType={detailData.category}
+            placeName={detailData.name}
+            latitude={Number(detailData.latitude)}
+            longitude={Number(detailData.longitude)}
+          />
           <PlaceSlider images={detailData.imageUrls} name={detailData.name} />
-
-          <RelevantContainer>
-            <RelevantHeader>
-              <Title size="sm">이 장소가 포함된 코스</Title>
-              <Link href={`/course/search/${placeId}`}>
-                <Text color="gray">전체보기</Text>
-              </Link>
-            </RelevantHeader>
-            <RelevantCourses courses={relevantCourses} />
-          </RelevantContainer>
+          {relevantCourses.length > 0 && (
+            <RelevantCourses placeId={placeId} courses={relevantCourses} />
+          )}
           <Comment id={detailData.id} type="place" />
           <HorizonDivideLine />
           <DetailSidebar
@@ -150,34 +128,8 @@ const Container = styled(PageContainer)`
   margin-top: 40px;
 `;
 
-const PostHeader = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 6px;
-`;
-
 const HorizonDivideLine = styled.div`
   width: 100%;
   height: 1px;
   background-color: ${theme.color.backgroundDarkGray};
-`;
-
-const MapButton = styled(Button)`
-  margin-top: 4px;
-`;
-
-const RelevantContainer = styled.div`
-  margin-top: 30px;
-  padding: 50px 0;
-  border-top: 1px solid ${theme.color.borderGray};
-`;
-
-const RelevantHeader = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-`;
-
-const Description = styled(Text)`
-  margin-top: 4px;
 `;
